requirements.txt
Django>=4.2
qrcode[pil]>=7.4
Pillow>=10.0
user-agents>=2.2
